move communicate() and unicode_decode() to utils completely
